% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AF}
\alias{AF}
\title{Example allele frequencies}
\format{
A numeric vector of length 19490.
}
\usage{
AF
}
\description{
A vector of allele frequencies for variants on chromosome 19,
provided for use in examples and vignettes.
}
\keyword{datasets}
