% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{direct_to_total}
\alias{direct_to_total}
\alias{total_to_direct}
\title{Solves for total/ effects from direct effects}
\usage{
direct_to_total(G_dir, restrict_dag = TRUE)

total_to_direct(G_tot, restrict_dag = TRUE)
}
\arguments{
\item{G_dir}{Direct effect adjacency matrix}

\item{restrict_dag}{Should the function fail if G_tot is not a DAG?}

\item{G_tot}{Total effect adjacency matrix}
}
\value{
Total effect adjacency matrix

Direct effect adjacency matrix
}
\description{
Solves for total/ effects from direct effects
}
\examples{
X <- structure(c(0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0), dim = c(4L,4L))
print(all(total_to_direct(direct_to_total(X)) == X))
}
