% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size_distribution.R
\name{effect_dist_traits}
\alias{effect_dist_traits}
\title{List all available trait names}
\usage{
effect_dist_traits()
}
\value{
Character vector.
}
\description{
Returns the vector of trait names available in [effect_dist()].
}
\examples{
effect_dist_traits()
}
