% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_eigen.R
\name{fast_eigen}
\alias{fast_eigen}
\title{Fast Eigen Decomposition using RcppArmadillo}
\usage{
fast_eigen(m)
}
\arguments{
\item{m}{A symmetric numeric matrix.}
}
\value{
A list with two components: `values` (eigenvalues) and `vectors` (eigenvectors).
}
\description{
Computes the full eigen decomposition of a symmetric matrix using C++.
This function is faster than base R's `eigen()`.
This function will not check that your matrix is numeric or symmetric.
}
