% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_eigen.R
\name{fast_eigen_vals}
\alias{fast_eigen_vals}
\title{Fast Eigenvalues using RcppArmadillo}
\usage{
fast_eigen_vals(m)
}
\arguments{
\item{m}{A symmetric numeric matrix.}
}
\value{
A vector of eigen values
}
\description{
Computes eigen values only of a symmetric matrix using C++.
This function will not check that your matrix is numeric or symmetric.
}
