% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_inddata.R
\name{fast_lm}
\alias{fast_lm}
\title{Fast per-variant marginal linear regression}
\usage{
fast_lm(X, Y, check = TRUE)
}
\arguments{
\item{X}{Genotype matrix (individuals by variants).}

\item{Y}{Phenotype matrix (individuals by traits). NAs are allowed.}

\item{check}{If TRUE, validate \code{X} and \code{Y}.}
}
\value{
A data frame with columns \code{bhat_m} and \code{s_m} for each trait \code{m}.
}
\description{
Fast per-variant marginal linear regression
}
\keyword{internal}
