% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixnorm_to_scalefam.R
\name{fixed_to_scale_fam}
\alias{fixed_to_scale_fam}
\title{Build a scale-family effect function from a fixed effect vector}
\usage{
fixed_to_scale_fam(b)
}
\arguments{
\item{b}{Vector of fixed effects.}
}
\value{
A function with arguments \code{n} and \code{sd} that returns rescaled effects.
}
\description{
Build a scale-family effect function from a fixed effect vector
}
\keyword{internal}
