% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_g.R
\name{G_from_text}
\alias{G_from_text}
\title{Generate G matrix from text input}
\usage{
G_from_text(text)
}
\arguments{
\item{text}{A multi line text string that specifies the causal relationships between traits}
}
\value{
A square matrix with dimension number of traits
}
\description{
Convenience function for parsing a multi-line text string into the G matrix, where each line 
specifies one causal relationship between traits. Each line must be formatted with four 
space-separated tokens as:
}
\details{
<trait 1> <causal orientation> <trait 2> <effect>

The function tries to parse the effect from numeric expressions or global variables.
}
\examples{

# Simple specification
text <- "X <- Y 0.25
         A <- X 0.1
         A <- Y sqrt(0.34)"
G_from_text(text)

# Another example
a <- 0.25
b <- -0.34
text <- "X -> Y a
         A -> X b
         Y <- A sqrt(0.34)"
G_from_text(text)
}
