% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_bhat_from_b.R
\name{gen_bhat_from_b}
\alias{gen_bhat_from_b}
\title{Generate beta hats from standardized direct SNP effects and LD}
\usage{
gen_bhat_from_b(
  b_joint_std,
  b_joint,
  trait_corr,
  N,
  R_LD = NULL,
  snp_info = NULL,
  af = NULL,
  L_mat_joint_std = NULL,
  theta_joint_std = NULL
)
}
\arguments{
\item{b_joint_std}{variants by traits matrix of direct joint SNP effects.}
}
\description{
Generate beta hats from standardized direct SNP effects and LD
}
