% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_inddata.R
\name{gen_genos_mvn}
\alias{gen_genos_mvn}
\title{Simulate genotypes via a multivariate normal approximation to LD}
\usage{
gen_genos_mvn(n, J, R_LD, af)
}
\arguments{
\item{n}{Number of individuals.}

\item{J}{Number of variants.}

\item{R_LD}{LD pattern (optional). See \code{?sim_mv} for more details.}

\item{af}{Allele frequencies.}
}
\value{
A list with elements \code{X} (a genotype matrix) and \code{af}.
}
\description{
Simulate genotypes via a multivariate normal approximation to LD
}
\keyword{internal}
