% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_F.R
\name{generate_F2}
\alias{generate_F2}
\title{Generate a random factor matrix with specified sparsity and row sums}
\usage{
generate_F2(
  non_zero_by_factor,
  square_row_sums,
  rfunc = function(n) {
     stats::runif(n, -1, 1)
 }
)
}
\arguments{
\item{non_zero_by_factor}{Vector giving the number of non-zero elements per factor (columns).}

\item{square_row_sums}{Vector of target row sums of squared entries.}

\item{rfunc}{Function used to draw non-zero entries.}
}
\value{
A factor matrix \code{F_mat} with the requested sparsity and rescaled rows.
}
\description{
Generate a random factor matrix with specified sparsity and row sums
}
\keyword{internal}
