% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_F.R
\name{generate_F_simple}
\alias{generate_F_simple}
\title{Generate simple block factor matrices}
\usage{
generate_F_simple(
  nblocks,
  type = c("nested", "difference", "checkers1", "checkers2")
)
}
\arguments{
\item{nblocks}{Number of blocks.}

\item{type}{Type of pattern: one of "nested", "difference", "checkers1", or "checkers2".}
}
\value{
A factor matrix \code{F_mat}.
}
\description{
Generate simple block factor matrices
}
\keyword{internal}
