% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_F.R
\name{generate_random_F}
\alias{generate_random_F}
\title{Generate random F}
\usage{
generate_random_F(
  K,
  M,
  g_F = function(n) {
     runif(n, -1, 1)
 },
  nz_factor,
  omega,
  h2_trait,
  pad = FALSE
)
}
\arguments{
\item{K}{K}

\item{M}{M}

\item{g_F}{Function from which non-zero elements of F are generated}

\item{nz_factor}{Number of non-zero elements of each factor if F is to be generated.}

\item{omega}{Proportion of trait heritability explained by factors}

\item{h2_trait}{Trait heritability}

\item{pad}{Add single trait factors? (See details)}
}
\value{
A matrix
}
\description{
Generate random F
}
\details{
Generate a random set of (at least) K factors for M traits.
The number of traits affected by each factor is given by \code{nz_factor}.
Each effect is chosen as a random draw from function g_F.
If any rows of the resulting matrix corresponding to non-zero elements of omega
are all zero and pad  = TRUE, single-trait factors are added.
Finally, the
matrix is re-scaled so that \code{colSums(F_mat^2) = omega*h2_trait}.
}
