% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ld_mat_list}
\alias{ld_mat_list}
\title{Example LD block matrices}
\format{
A list of 39 sparse correlation matrices (class \code{dsCMatrix}).
}
\usage{
ld_mat_list
}
\description{
A list of LD block correlation matrices used to demonstrate
simulation with LD.
}
\keyword{datasets}
