% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixnorm_to_scalefam.R
\name{mixnorm_to_scale_fam}
\alias{mixnorm_to_scale_fam}
\title{Build a scale-family effect function from a normal mixture}
\usage{
mixnorm_to_scale_fam(sigma, pi)
}
\arguments{
\item{sigma}{Vector of component standard deviations.}

\item{pi}{Vector of mixture proportions, summing to 1.}
}
\value{
A function with arguments \code{n} and \code{sd} for generating effects.
}
\description{
Build a scale-family effect function from a normal mixture
}
\keyword{internal}
