% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mv.R
\name{print.sim_mv}
\alias{print.sim_mv}
\title{Print sim_mv}
\usage{
\method{print}{sim_mv}(x, ...)
}
\arguments{
\item{x}{A sim_mv object}

\item{...}{Arguments to be passed to str}
}
\description{
Print sim_mv
}
