% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hapsim_simple.R
\name{R_LD_to_haplodat}
\alias{R_LD_to_haplodat}
\title{Convert LD matrices to haplotype data}
\usage{
R_LD_to_haplodat(R_LD, af)
}
\arguments{
\item{R_LD}{A list of LD matrices, each matrix corresponding to a set of loci.}

\item{af}{A vector of allele frequencies corresponding to the loci in `R_LD`.}
}
\value{
A list of haplotype data, where each element corresponds to a set of loci.
}
\description{
Converts a list of LD matrices to haplotype data.
}
\details{
The function formats LD matrices into hapsim formatted input data.
}
