% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_inddata.R
\name{resample_inddata}
\alias{resample_inddata}
\title{Sample individual level data with joint effects matching a sim_mv object}
\usage{
resample_inddata(
  N,
  dat = NULL,
  genos = NULL,
  J = NULL,
  R_LD = NULL,
  af = NULL,
  sim_func = gen_genos_mvn,
  new_env_var = NULL,
  new_R_E = NULL,
  new_R_obs = NULL,
  calc_sumstats = FALSE
)
}
\arguments{
\item{N}{Sample size, scalar, vector, or special sample size format data frame, see details.}

\item{dat}{An object of class \code{sim_mv} (produced by \code{sim_mv}). If `dat` is omitted,
the function will generate a matrix of genotypes only. If `dat` is provided, phenotypes for the traits
in `dat` will also be included.}

\item{genos}{Optional matrix of pre-generated genotypes. If \code{genos} is supplied, \code{resample_inddata} will only generate phenotypes.}

\item{J}{Optional number of variants. \code{J} is only required if \code{dat} is missing.}

\item{R_LD}{LD pattern (optional). See \code{?sim_mv} for more details.}

\item{af}{Allele frequencies. \code{af} is required unless unless \code{genos} is supplied.}

\item{new_env_var}{Optional. The environmental variance in the new population.
If missing the function will assume the environmental variance is the same as in the old population.}

\item{new_R_E}{Optional, specify environmental correlation in the new population.
If missing, the function will assume the environmental correlation is the same as in the original data.}

\item{new_R_obs}{Optional, specify overall trait correlation in the new population. Specify at most one of \code{new_R_E} or \code{new_R_obs}.
If missing, the function will assume the environmental correlation is the same as in the original data.}

\item{calc_sumstats}{If \code{TRUE}, associations between genotypes and phenotypes will be calculated and returned.}
}
\description{
Sample individual level data with joint effects matching a sim_mv object
}
\details{
This function can be used to generate individual level genotype and phenotype data. It can be used in three modes:

To generate genotype data only: No \code{sim_mv} object needs to be included. Supply only \code{N} as a single integer for the
number of individuals, \code{J} for the number of variants, \code{af}, and \code{R_LD} if desired. All other
parameters are not relevant if there is no phenotype, so if they are supplied, you will get an error. The returned object will include a
\code{N x J} matrix of genotypes and a vector of allele frequencies.

To generate both genotype and phenotype data: Supply \code{dat} (a \code{sim_mv} object) and leave \code{genos} missing. \code{N} and \code{af}
are required and all other options are optional.

To generate phenotype data only: Supply \code{dat} (a \code{sim_mv} object) and provide a matrix of genotypes to the \code{genos} argument. \code{N}
must be consistent with the number of genotypes supplied. So for example, if there are two traits with 10 samples each and no overlap, \code{genos}
should have 20 rows. \code{N} and \code{af}
are required and all other options are optional.
}
\examples{
# Use gen_gwas_from_b to generate individual level data with given effect size.
Ndf <- data.frame(trait_1 = 1, trait_2 = 1, N = 10000)
G <- matrix(0, nrow = 2, ncol = 2)
R_E <- matrix(c(1, 0.8, 0.8, 1), nrow = 2, ncol = 2)
# original data
dat <- sim_mv(N = Ndf, J = 2000, h2 = c(0.4, 0.3), pi = 100/2000,
               G = G, R_E = R_E, af = function(n){rbeta(n, 1, 5)})
# Now generate GWAS data
gw_dat <- resample_inddata(dat, N = Ndf, calc_sumstats = TRUE)
}
