% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_sumstats.R
\name{resample_sumstats}
\alias{resample_sumstats}
\title{Resample Summary Statistics for Existing Simulation Object}
\usage{
resample_sumstats(
  dat,
  N,
  R_LD = NULL,
  af = NULL,
  est_s = FALSE,
  geno_scale = c("allele", "sd"),
  new_env_var = NULL,
  new_R_E = NULL,
  new_R_obs = NULL
)
}
\arguments{
\item{dat}{Object output by \code{sim_mv}}

\item{N}{Sample size, scalar, vector, matrix. See \code{?sim_mv} for more details.}

\item{R_LD}{LD pattern (optional). See \code{?sim_mv} for more details.}

\item{af}{Allele frequencies. See \code{?sim_mv} for more details.}

\item{est_s}{Logical, should estimates of se(beta_hat) be produced.}

\item{geno_scale}{Either "allele" or "sd". Specifies the scale of the effect sizes in the output data.}

\item{new_env_var}{Optional. The environmental variance in the new population.
If missing the function will assume the environmental variance is the same as in the old population.}

\item{new_R_E}{Optional, specify environmental correlation in the new population.
If missing, the function will assume the environmental correlation is the same as in the original data.}

\item{new_R_obs}{Optional, specify overall trait correlation in the new population. Specify at most one of \code{new_R_E} or \code{new_R_obs}.
If missing, the function will assume the environmental correlation is the same as in the original data.}
}
\description{
Resample Summary Statistics for Existing Simulation Object
}
\details{
This function can be used to generate new summary statistics for an existing simulation object.
}
\examples{
# Use resample_sumstats to generate new GWAS results with the same effect sizes.
N <- matrix(1000, nrow = 2, ncol =2)
G <- matrix(0, nrow = 2, ncol = 2)
R_E <- matrix(c(1, 0.8, 0.8, 1), nrow = 2, ncol = 2)
# original data
dat <- sim_mv(N = N, J = 20000, h2 = c(0.4, 0.3), pi = 1000/20000,
               G = G, R_E = R_E)
# data for second GWAS
# Since we didn't supply af or an LD pattern in the original GWAS,
# we have standardized effects.
dat_new <- resample_sumstats(dat, N = 40000)
}
