% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixnorm_to_scalefam.R
\name{rnormalmix}
\alias{rnormalmix}
\title{Simulate from a normal mixture distribution}
\usage{
rnormalmix(n, sd, pi, mu = 0, return.Z = FALSE)
}
\arguments{
\item{n}{Number of points to simulate}

\item{sd}{Standard deviations}

\item{pi}{Mixture proportions}

\item{mu}{Means}

\item{return.Z}{if TRUE, also return a vector of indicators indicating which of the K classes each sample belongs to}
}
\value{
If return.Z=TRUE, returns a list with elements beta (samples) and Z (indicators). Otherwise returns a length n vector of samples.
}
\description{
Simulate from a normal mixture distribution
}
