% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ld_prune.R
\name{sim_extract_ld}
\alias{sim_extract_ld}
\title{Extract LD matrix from simulated data}
\usage{
sim_extract_ld(dat, index, R_LD)
}
\arguments{
\item{dat}{Simulation object produced by `sim_mv`}

\item{index}{vector of indices for snps to extract LD}

\item{R_LD}{List of eigen-decompositions used in original simulation}
}
\value{
An LD matrix. SNP order matches original index order
}
\description{
Extract LD matrix for specific variants in simulated data set
}
\examples{
data("ld_mat_list")
data("snpdata")
# Two traits with no causal relationship, non-overlapping GWAS
set.seed(1)
G <- matrix(0, nrow = 2, ncol = 2)
dat <- sim_mv(N = 10000, J = 50000, h2 = c(0.4, 0.3), pi = 1000/20000,
               G = G,  R_LD = ld_mat_list, snp_info = snpdata)

# extract ld matrix for all variants with p-value for trait 1 less than 1e-5
pvals <- 2*pnorm(-abs(dat$beta_hat/dat$se_beta_hat))
index <- which(pvals[,1] < 1e-5)
ld_mat <- sim_extract_ld(dat, index, ld_mat_list)
dim(ld_mat)
length(index)
}
