% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ld_prune.R
\name{sim_ld_proxy}
\alias{sim_ld_proxy}
\title{Retrieve LD proxies}
\usage{
sim_ld_proxy(dat, index, R_LD, r2_thresh = 0.64, return_mat = FALSE)
}
\arguments{
\item{dat}{Simulation object produced by `sim_mv`}

\item{index}{list of indexes to retrieve proxies for}

\item{R_LD}{LD pattern used to generate dat}

\item{r2_thresh}{Get proxies with r^2 >= r2_thresh with one of the index variants}

\item{return_mat}{If TRUE, return the correlation matrix between the index variant and the proxies. In this matrix, the
the index variant always corresponds the first row/column and the proxies are in the order returned.}
}
\description{
Retrieve LD proxies
}
