% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ld_prune.R
\name{sim_ld_prune}
\alias{sim_ld_prune}
\title{LD prune simulated data}
\usage{
sim_ld_prune(dat, pvalue, R_LD, r2_thresh = 0.1, pval_thresh = 1)
}
\arguments{
\item{dat}{Data object produced by sim_mv}

\item{pvalue}{A vector of p-values to use to prioritize variants.}

\item{R_LD}{The same list of eigen-decompositions used to generate dat}

\item{r2_thresh}{r^2 thershold for pruning}

\item{pval_thresh}{p-value threshod for pruning (see details)}
}
\value{
A vector of indices corresponding to the LD-pruned variant set.
}
\description{
LD prune simulated data
}
\details{
Given results from sim_mv, and a vector of p-values,
the function will return a list of variants that have \code{p < pval_thresh}
and which mutually have squared correlation less than \code{r2_thresh}.
}
\examples{
data("ld_evd_list")
data("snpdata")
head(snpdata)
# Two traits with no causal relationship, non-overlapping GWAS
set.seed(1)
G <- matrix(0, nrow = 2, ncol = 2)
dat <- sim_mv(N = 10000, J = 50000, h2 = c(0.4, 0.3), pi = 1000/20000,
               G = G,  R_LD = ld_evd_list, snp_info = snpdata)

# prune on p-value for first trait
pvals <- 2*pnorm(-abs(dat$beta_hat/dat$se_beta_hat))
prune_set_1 <- sim_ld_prune(dat, pvalue = pvals[,1], R_LD = ld_evd_list, pval_thresh = 1e-5)
length(prune_set_1)
}
