% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R, R/simGWAS-package.R
\docType{package}
\name{simGWAS}
\alias{simGWAS}
\title{simGWAS}
\description{
simGWAS

simGWAS: Simulate GWAS summary statistics from specified DAG or factor structure.
}
\author{
Jean Morrison <jvmorr@umich.edu>
}
