% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snpdata}
\alias{snpdata}
\title{Example SNP data}
\format{
A data frame with 19490 rows and 14 variables including allele frequency,
rs ID, chromosome, position, LD block id, and LD pruning indicators.
}
\usage{
snpdata
}
\description{
Variant information accompanying \code{\link{AF}} and \code{\link{ld_mat_list}},
provided for use in examples and vignettes.
}
\keyword{datasets}
