% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyz_to_G.R
\name{xyz_to_G}
\alias{xyz_to_G}
\title{Generate G from XYZ Specification}
\usage{
xyz_to_G(tau_xz, tau_yz, dir_xz, dir_yz, beta)
}
\arguments{
\item{dir_xz, dir_yz}{Effect direction between Z and X or Y (see details)}

\item{beta}{Signed variance of Y explained by X, see details}

\item{taux_xz, tau_yz}{Effect size between Z and X or Y as signed percent variance explained (see details)}
}
\value{
A matrix of direct effects corresponding to variables in the order (Y, X, Z_1, ..., Z_K)
}
\description{
Generate G from XYZ Specification
}
\details{
This function generates a matrix G of direct effects corresponding to a model in with variables Y, X, and Z_1, ..., Z_K.
There is a causal effect of X on Y given by beta, which specifies the proportion of the variance of Y explained by X.
The K variables Z_1, ..., Z_K can have
effects to or from X and Y but do not have direct effects on each other. Vectors \code{dir_xz} and \code{dir_yz} specify the direction
of these effects with +1 corresponding to "to" effects and -1 corresponding to "from" effects. For example, \code{dir_xz = c(1, -1)}
and \code{dir_yz = c(1, 1)} would indicate two variables Z_1 and Z_2 with Z_1 being a common cause of both X and Y and Z_2
being a cause of $Y$ and caused by X (Z_2 is a mediator between X and Y). The function will give an error if there
is any index with \code{dir_xz} equal to -1 and \code{dir_yz} equal 1. This would indicate a variable that is a mediator between Y and
X, however, because there is an effect from X to Y assumed, the resulting graph would be cyclic and not allowed.

The inputs `tau_xz` and `tau_yz` specify the effect sizes of each of the Z_k variables on or
from X and $Y$. These are given in signed percent variance explained. If we again used \code{dir_xz = c(1, -1)} and \code{dir_yz = c(1, 1)}
with \code{tau_xz = c(0.2, -0.3)} and \code{tau_yz = c(0.1, 0.25)},
this means that the confounder, Z_1 explains 20\% of the variance of X and 10\% of the
variance of Y and both effects are positive. X explains 30\% of the variance of the mediate Z_2 with a negative effect direction and
Z_2 explains 25\% of the variance of Y.
}
\examples{
xyz_to_G(tau_xz = c(0.2, -0.3), tau_yz = c(0.1, 0.25),
        dir_xz = c(1, -1), dir_yz = c(1,1), beta = 0)
# code below will give an error due to specification of a cyclic graph.
xyz_to_G(tau_xz = c(0.2, -0.3), tau_yz = c(0.1, 0.25),
        dir_xz = c(1, -1), dir_yz = c(-1,1), beta = 0.1)
# with beta = 0, there is no cycle so no error
xyz_to_G(tau_xz = c(0.2, -0.3), tau_yz = c(0.1, 0.25),
        dir_xz = c(1, -1), dir_yz = c(-1,1), beta = 0)
}
