% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simGWAS.R
\docType{data}
\name{AF}
\alias{AF}
\title{Allele frequencies for SNP data}
\format{
An object of class \code{numeric} of length 19490.
}
\usage{
AF
}
\description{
Allele frequency column from `snpdata` object
}
\keyword{datasets}
