% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_inddata.R
\name{fast_lm}
\alias{fast_lm}
\title{Fast implementation of linear model}
\usage{
fast_lm(X, Y, check = TRUE)
}
\arguments{
\item{X}{X vector or matrix}

\item{Y}{Y vector or matrix}

\item{check}{X and Y inputs}
}
\description{
Fast implementation of linear model
}
