% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_bhat_from_b.R
\name{gen_bhat_from_b}
\alias{gen_bhat_from_b}
\title{Generate beta hats from standardized or non-standardized direct SNP effects and LD}
\usage{
gen_bhat_from_b(
  b_joint,
  N,
  trait_corr = NULL,
  R_LD = NULL,
  R_LD_eigen = NULL,
  af = NULL,
  est_s = FALSE,
  input_geno_scale = c("allele", "sd"),
  input_pheno_sd = 1,
  output_geno_scale = c("allele", "sd"),
  output_pheno_sd = 1
)
}
\arguments{
\item{b_joint}{Matrix of joint effect sizes}

\item{N}{Sample size, scalar, vector, matrix, or data.frame. See \code{?sim_mv} for more details.}

\item{trait_corr}{Matrix of population trait correlation (traits by traits)}

\item{R_LD}{LD pattern (optional). See \code{?sim_mv} for more details.}

\item{R_LD_eigen}{Optionally, also supply list of eigen decompositions. The program will not check that this matches R_LD.}

\item{af}{Allele frequencies (optional, allowed only if \code{R_LD} is missing). See \code{?sim_mv} for more details.}

\item{est_s}{Estimate standard errors?}

\item{input_geno_scale}{Genotype scale of effects in  b_joint}

\item{input_pheno_sd}{Phenotype sd for effects in b_joint}

\item{output_geno_scale}{Output genotype scale}

\item{output_pheno_sd}{Output phenotype sd}
}
\description{
Generate beta hats from standardized or non-standardized direct SNP effects and LD
}
\details{
This has been made an internal function. To resample summary statistics use \code{resample_sumstats}.
}
