% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_inddata.R
\name{gen_genos_mvn}
\alias{gen_genos_mvn}
\title{Generate genotypes from multivariate normal}
\usage{
gen_genos_mvn(n, J, R_LD, af)
}
\arguments{
\item{n}{Sample size to generate}

\item{J}{Number of variants to generate}

\item{R_LD}{Optional list of LD blocks. R_LD should have class \code{list}.
Each element of R_LD can be either a) a matrix, b) a sparse matrix (class \code{dsCMatrix}) or c) an eigen decomposition (class \code{eigen}).
All elements should be correlation matrices, meaning that they have 1 on the diagonal and are positive definite. See Details and vignettes.}

\item{af}{Optional vector of allele frequencies. If R_LD is not supplied, af can be a scalar, vector or function.
If af is a function it should take a single argument (n) and return a vector of n allele frequencies (See Examples).
If R_LD is supplied, af must be a vector with length equal to the size of the supplied LD pattern (See Examples).}
}
\description{
Generate genotypes from multivariate normal
}
