% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_F.R
\name{generate_F2}
\alias{generate_F2}
\title{Generate F (2)}
\usage{
generate_F2(
  non_zero_by_factor,
  square_row_sums,
  rfunc = function(n) {
     runif(n, -1, 1)
 }
)
}
\arguments{
\item{non_zero_by_factor}{non_zero_by_factor}

\item{rfunc}{rfunc}
}
\description{
Generate F (2)
}
