% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_F.R
\name{generate_F_simple}
\alias{generate_F_simple}
\title{Generate F (simple)}
\usage{
generate_F_simple(
  nblocks,
  type = c("nested", "difference", "checkers1", "checkers2")
)
}
\arguments{
\item{nblocks}{nblocks}

\item{type}{type}
}
\description{
Generate F (simple)
}
