% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simGWAS.R
\docType{data}
\name{ld_mat_list}
\alias{ld_mat_list}
\title{Block LD matrices for SNP data}
\format{
An object of class \code{list} of length 39.
}
\usage{
ld_mat_list
}
\description{
Block LD matrices for SNP data
}
\keyword{datasets}
