% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixnorm_to_scalefam.R
\name{mixnorm_to_scale_fam}
\alias{mixnorm_to_scale_fam}
\title{mixnorm_to_scale_fam}
\usage{
mixnorm_to_scale_fam(sigma, pi)
}
\arguments{
\item{sigma}{sigma}

\item{pi}{pi}
}
\description{
mixnorm_to_scale_fam
}
