% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_sumstats.R
\name{resample_sumstats}
\alias{resample_sumstats}
\title{Resample Summary Statistics for Existing Simulation Object}
\usage{
resample_sumstats(sim_dat, N, R_LD = NULL, af = NULL, est_s = FALSE)
}
\arguments{
\item{sim_dat}{Object output by \code{sim_mv}}

\item{N}{Sample size, scalar, vector, matrix. See \code{?sim_mv} for more details.}

\item{R_LD}{LD pattern (optional). See \code{?sim_mv} for more details.}

\item{af}{Allele frequencies (optional, allowed only if \code{R_LD} is missing). See \code{?sim_mv} for more details.}
}
\description{
Resample Summary Statistics for Existing Simulation Object
}
\details{
This function can be used to generate new summary statistics for an existing simulation object.
The new summary statistics will have the same true causal effects as the original. However, you can use a new
sample size and new LD if desired. This function is primarily a wrapper for \code{gen_bhat_from_b}. This function
differs from \code{gen_bhat_from_b} in its arguments. It can accept the original simulated data directly rather than
requiring the  user to extract and supply the joint effects. It will also preserver the direct and total trait effects objects in
the original data.

Note about effect scalings: If the original simulation object was generated with no allele frequencies, then
all effect size and effect estimate related outputs will all be on the standardized scale (in units of SD change in Y per sd change in genotype).
If you use this object in \code{resample_sumstats} but add allele frequencies, the resulting object will be converted
to non-standardized effects (effects in units of SD change in Y per alternate allele).

If the original simulation was generated with allele frequencies, then the original object will be on the non-standardized scale.
\code{resample_sumstats} will sample summary statistics assuming the same non-standardized effects. This means that if you use
different allele frequencies with \code{resample_sumstats} than used with the original, the effects in the two objects will differ
on the standardized scale. The rationale for this behavior is that the non-standardized effects are the "biological" effects.

Note that changing the allele frequency and/or LD will lead to a change in hertiability and genetic covariance (\code{Sigma_G}).
}
\examples{
# Use resample_sumstats to generate new GWAS results with the same effect sizes.
N <- matrix(1000, nrow = 2, ncol =2)
G <- matrix(0, nrow = 2, ncol = 2)
R_E <- matrix(c(1, 0.8, 0.8, 1), nrow = 2, ncol = 2)
# original data
dat <- sim_mv(N = N, J = 20000, h2 = c(0.4, 0.3), pi = 1000/20000,
               G = G, R_E = R_E)
# data for second GWAS
# Since we didn't supply af or an LD pattern in the original GWAS,
# we have standardized effects.
dat_new <- resample_sumstats(dat, N = 40000)
}
