% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ld_prune.R
\name{sim_ld_proxy}
\alias{sim_ld_proxy}
\title{Retrieve LD proxies}
\usage{
sim_ld_proxy(snpinfo, index, R_LD, r2_thresh = 0.64)
}
\arguments{
\item{snpinfo}{snp_info data frame from a simulation object produced by sim_mv}

\item{index}{list of indexes to retrieve proxies for}

\item{R_LD}{list of eigen-decompositions used originally in sim_mv}

\item{r2_thresh}{Get proxies with r^2 >= r2_thresh with one of the index variants}
}
\description{
Retrieve LD proxies
}
