% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_sumstats_lf.R
\name{sim_sumstats_lf}
\alias{sim_sumstats_lf}
\title{Simulate summary stats}
\usage{
sim_sumstats_lf(
  F_mat,
  N,
  J,
  h_2_trait,
  omega,
  h_2_factor,
  pi_L,
  pi_theta,
  R_E,
  maf = NA,
  R_LD = NULL,
  snp_info = NULL,
  g_F,
  nz_factor,
  overlap_prop = 1,
  sporadic_pleiotropy = TRUE
)
}
\arguments{
\item{F_mat}{factor matrix M by K}

\item{N}{GWAS sample size}

\item{J}{Total number of SNPs to generate}

\item{h_2_trait}{Heritability of each trait. Length M vector.}

\item{omega}{Proportion of SNP heritability mediated by factors. Length M vector.}

\item{h_2_factor}{Heritability of each factor. Length K vector.}

\item{pi_L}{Proportion of non-zero elements in L_k. Length K factor}

\item{pi_theta}{Proportion of non-zero elements in theta. Scalar or length M vector.}

\item{R_E}{Environmental trait correlation not mediated by factors. M by M pd matrix.}

\item{maf}{can either be a scalar in which case the same maf is used for all SNPS, NA in which case SNPs
are assumed scaled to variance 1, a function that takes a number n and returns n values between 0 and 1, or a
vector of length equal to the number of SNPs. maf is ignored if LD is provided.}

\item{R_LD}{List of eigen decompositions of LD correlation matrices, may be missing.}

\item{snp_info}{If R_LD is provided, provide a data frame with columns "SNP" and "AF"}

\item{g_F}{Function from which non-zero elements of F are generated}

\item{nz_factor}{Number of non-zero elements of each factor if F is to be generated.}

\item{sporadic_pleiotropy}{Allow a single SNP to affect multiple factors (default TRUE).}
}
\description{
Simulate summary statistcs for fully overlapping GWAS with no LD
}
\details{
If F_mat is not provided, it will be generated using the `generate_F2` function.
In this case g_F and nz_factor must be provided. All of the elements
in F are generated iid from a mixture of a point mass at 0 and g_F. The matrix is then
rescaled according to the constraints.


With this setup it is possible to specify a setting that is impossible. Usually this occurs if the heritability
of the factors is low but the heritability of the traits is high leading to a contradiction. Right now the function
will just return an error if that happens.
}
